% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{biom2generic}
\alias{biom2generic}
\title{Converts biom-object}
\usage{
biom2generic(biom.file)
}
\arguments{
\item{biom.file}{Name of a biom-file output from DEBLUR,}

\item{out.folder}{Name of folder to output the resulting files.}
}
\value{
The function will produce two files: The file named
\code{readcount_table.txt} containing the tab-separated tale of readcounts,
one row for each OTU and and column for each sample. The first column is
the names of the OTU's. The second file is named \code{centroids.fasta} and
contains the OTU-sequences, i.e. it has the same number of entries as there
are rows in the \code{readcount_table.txt}.
}
\description{
The content of a biom-object is converted into a readcount-table
and a fasta-object and written to files.
}
\details{
This function is used in our DEBLUR pipeline, to convert the
output from the biom-file to more generic files containing a readcount-
table and sequences in fasta format.
}
\author{
Lars Snipen.
}
