% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiv_wrangling.R
\name{filterOTU}
\alias{filterOTU}
\title{Filter OTUs}
\usage{
filterOTU(midiv.obj, OTU_to_keep)
}
\arguments{
\item{midiv.obj}{A midiv object (list), see \code{\link{midivObject}}.}

\item{OTU_to_keep}{A logical vector indicating which OTUs to keep (TRUE).}
}
\value{
A midiv object (list) with (potentially) fewer OTUs.
}
\description{
Filtering OTUs in a midiv object.
}
\details{
This function filters OTUs, i.e. keep a subsets of rows in the
\code{readcount.mat} and \code{sequence.tbl} inside the \code{midiv.obj} list.

The \code{OTU_to_keep} should be a logical vector with one element for each
OTU in the object.
}
\author{
Lars Snipen.
}
