% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiv_wrangling.R
\name{filterSample}
\alias{filterSample}
\title{Filter samples}
\usage{
filterSample(midiv.obj, samples_to_keep)
}
\arguments{
\item{midiv.obj}{A midiv object (list), see \code{\link{midivObject}}.}

\item{samples_to_keep}{A logical vector indicating which sampless to keep (TRUE).}
}
\value{
A midiv object (list) with (potentially) fewer samples.
}
\description{
Filtering samples in a midiv object.
}
\details{
This function filters samples, i.e. keep a subsets of rows in the
metadata.tbl and corresponding columns in the \code{readcount.mat} of the
\code{midiv.obj} list.

The \code{samples_to_keep} should be a logical vector with one element for each
sample in the object.
}
\author{
Lars Snipen.
}
