% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAtesting.R
\name{nonparDA}
\alias{nonparDA}
\title{Non-parametric DA testing}
\usage{
nonparDA(ps.obj)
}
\arguments{
\item{ps.obj}{A phyloseq object.}
}
\description{
Testing for differential abundance using non-parametric tests.
}
\author{
Lars Snipen.
}
