% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPS.R
\name{openPS}
\alias{openPS}
\title{Creating a phyloseq-like object}
\usage{
openPS(
  metadata.file,
  readcount.file,
  centroids.file,
  taxonomy.file = NULL,
  sintax.threshold = 0.8
)
}
\arguments{
\item{metadata.file}{Full name of text-file with metadata.}

\item{taxonomy.file}{Full name of text-file with the taxonomy table results (optional).}

\item{readcounts.file}{Full name of text-file with readcount data.}

\item{centroids.file.Full}{name of fasta-file with sequences.}
}
\value{
A list with the elements:

* sample_data.tbl - a data.frame with one row for each sample. Its rownames are the SampleID's.
* otu_table.mat - a matrix with the readcounts. These may later be transformed. The samples are in the columns, the OTUs in the rows.
* sequence.tbl - a data.frame with the OTU sequences (see \code{\link{readFasta}}).
* tax_table.mat - a matrix with the taxonomy, one row for each OTU.
}
\description{
Creating a an object similar to a phyloseq-object, but as an open list.
}
\details{
This function reads the input files, and store their content in a
\code{list} that we refer to as an *openPS object*.
}
\author{
Lars Snipen.
}
