% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPS.R
\name{openPS2phyloseq}
\alias{openPS2phyloseq}
\title{Convert to phyloseq object}
\usage{
openPS2phyloseq(openPS.obj)
}
\arguments{
\item{openPS.obj}{An openPS object, see \code{\link{openPS}}.}
}
\value{
A \code{\link{phyloseq}} object.
}
\description{
Creating a phyloseq object from an openPS object.
}
\details{
This function converts an openPS object, which is a simple
\code{list}, to a \code{\link{phyloseq}} object from the \code{phyloseq} R package.
}
\author{
Lars Snipen.
}
