% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiv_object.R
\name{phyloseq2midiv}
\alias{phyloseq2midiv}
\title{Convert phyloseq to midiv object}
\usage{
phyloseq2midiv(phyloseq.obj)
}
\arguments{
\item{phyloseq.obj}{A phyloseq object, see \code{\link{phyloseq}}.}
}
\value{
A \code{list} with entries as in a midiv object, except that the
\code{sequence.tbl} do not contain sequences, only taxonomy.
}
\description{
Creating a simple list from a phyloseq object.
}
\details{
This function converts a phyloseq object to a simple \code{\link{list}}
giving the entries the names as in \code{\link{midivObject}}.

This may be convenient for some wrangling on the data, and then perhaps converting
it back to a phyloseq object again with \code{\link{midiv2phyloseq}}.
}
\author{
Lars Snipen.
}
