% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPS.R
\name{phyloseq2openPS}
\alias{phyloseq2openPS}
\title{Convert to openPS object}
\usage{
phyloseq2openPS(phyloseq.obj)
}
\arguments{
\item{phyloseq.obj}{A phyloseq object, see \code{\link{phyloseq}}.}
}
\value{
A \code{list} with entries as in an openPS object.
}
\description{
Creating a simple list from a phyloseq object.
}
\details{
This function converts a phyloseq object to a simple \code{\link{list}}
giving the entries the names as in \code{\link{openPS}}.

This may be convenient for some wrangling on the data, and then perhaps converting
it back to a phyloseq object again with \code{\link{openPS2phyloseq}}.
}
\author{
Lars Snipen.
}
