% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sintax.R
\name{sintaxFilter}
\alias{sintaxFilter}
\title{Filter SINTAX taxonomy}
\usage{
sintaxFilter(taxonomy.tbl, sintax.threshold = 0.5)
}
\arguments{
\item{taxonomy.tbl}{A data.frame with SINTAX taxonomy results}

\item{sintax.threshold}{The threshold for accepting a classification}
}
\value{
A table with pure taxonomy results after filtering. The score columns
have been removed.
}
\description{
Filtering the taxonomy based on SINTAX scores.
}
\details{
A \code{taxonomy.tbl} from SINTAX has a score-column for each of the
ranks, e.g. the \code{domain} rank has a \code{domain_score}. If a score is below
the \code{sintax.threshold} the corresponding taxon is set to \code{NA}.

Finally, all score-columns are removed from the table.
}
\author{
Lars Snipen.
}
