% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{vsearch_update_metadata}
\alias{vsearch_update_metadata}
\title{Updating metadata file}
\usage{
vsearch_update_metadata(
  metadata.file,
  tmp.folder = "tmp_vsearch",
  readcounts.file.OTU = "readcounts_vsearch_OTU.txt",
  readcounts.file.ZOTU = "readcounts_vsearch_ZOTU.txt"
)
}
\arguments{
\item{metadata.file}{Full name of file with metadata (text file).}

\item{tmp.folder}{Name of folder where fastq files are found.}

\item{readcounts.file.OTU}{Name of file with OTU readcounts table.}

\item{readcounts.file.ZOTU}{Name of file with ZOTU readcounts table.}
}
\value{
The function will update the content of the file \code{metadata.file}.
}
\description{
Updating the metadata file with information on VSEARCH processing.
}
\details{
The \code{metadata.file} must be a text file with a table
with one row for each sample. It must follow the MiDiv metadata table standard
format. The columns used by this function are:
* ProjectID
* SequencingRunID
* SampleID

The names of the files in \code{tmp.folder} should all follow
the format: ProjectID_SequencingRunID_SampleID.fq.

This function will read the files and update the metadata file with the columns
n_vsearch_merged, n_vsearch_reads_OTU and n_vsearch_reads_ZOTU.
}
\author{
Lars Snipen.
}
