% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{De-multiplexing fastq files}
\usage{
demultiplex(
  sample.tbl,
  in.folder,
  out.folder,
  compress.out = TRUE,
  pattern = "_R[12].fastq"
)
}
\arguments{
\item{sample.tbl}{Table with data for each sample, see Details below.}

\item{in.folder}{Name of folder where raw fastq files are located.}

\item{out.folder}{Name of folder to output de-multiplexed fastq files.}

\item{compress.out}{Logical to indicate compressed output or not.}

\item{pattern}{The pattern to recognize the raw fastq files from other files}
}
\value{
The function will output the de-multiplexed fastq-files to the
\code{out.folder}. The name of each file consists of the corresponding
SampleID text, with the \code{_R1.fastq}/\code{_R2.fastq} extension. If
\code{compress.out = TRUE} the extension \code{.gz} is also added.

The function then returns these file extensions in R.
}
\description{
De-multiplexing Illumina data based on the extra forward barcode
used by MiDiv lab.
}
\details{
The input \code{sample.tbl} must be a table (tibble or data.frame)
with one row for each sample. It must follow the MiDiv sample_table standard
format. The columns used by this function are:
* SampleID
* Rawfile_R1
* Rawfile_R2
* Barcode

The SampleID should be a short text to uniquely identify each sample. It will
be used to form the new fastq filenames, so avoid using symbols not recommended
in filenames (e.g. space, slash). It is also wise to have a letter, not an integer,
as the first symbol in such SampleID texts.

Demultiplexing means extracting subsets of reads from raw fastq files, and saving
each subset on new fastq files. Each subset is identified by a barcode
sequence, and this must be listed in the Barcode column. The Barcode
sequence is matched at the start of the R1-reads only.

The original fastq files must all be in the \code{in.folder}. These files may
be compressed (.gz). The \code{pattern} is matched to all file names in this
folder, and should give a match against all raw fastq files. All new fastq
files written to \code{out.folder} will
also be compressed (.gz) if \code{compress.out = TRUE}.
}
\author{
Lars Snipen.
}
