% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiv_object.R
\name{midiv2phyloseq}
\alias{midiv2phyloseq}
\title{Convert midiv to phyloseq object}
\usage{
midiv2phyloseq(
  midiv.obj,
  sintax.threshold = 0.5,
  sample_id_column = "SampleID"
)
}
\arguments{
\item{midiv.obj}{A midiv object, see \code{\link{midivObject}}.}
}
\value{
A \code{\link{phyloseq}} object.
}
\description{
Creating a phyloseq object from a midiv object.
}
\details{
This function converts a midiv object, which is a simple
\code{list}, to a \code{\link{phyloseq}} object from the \code{phyloseq} R package.
}
\author{
Lars Snipen.
}
