% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiv_object.R
\name{midivObject}
\alias{midivObject}
\title{Creating a midiv-object}
\usage{
midivObject(
  metadata.tbl,
  readcount.mat,
  sequence.tbl,
  taxonomy.tbl = NULL,
  sample_id_column = "SampleID",
  filter_samples = "metadata",
  filter_OTUs = "sequence"
)
}
\arguments{
\item{metadata.tbl}{A data.frame with the sample metadata, or the name of a file.}

\item{readcount.mat}{A matrix with readcounts data, or the name of a file.}

\item{sequence.tbl}{A data.frame with the OTU sequences, or the name of a file.}

\item{taxonomy.tbl}{A data.frame with the OTU taxonomy table results, or the name of a file (optional).}

\item{sample_id_column}{Text with the name of the metadata.tbl column name that identifies samples.}

\item{filter_samples}{Text indicating if the metadata or the readcount table should decide the samples to keep.}

\item{filter_OTUs}{Text indicating if the readcount or sequence table should decide the OTUs to keep.}
}
\value{
A list with the elements:

\itemize{
  \item{\code{metadata.tbl}}{ a data.frame with one row for each sample.}
  \item{\code{readcount.mat}}{ a matrix with the readcounts, the samples are in the columns, the OTUs in the rows.}
  \item{\code{sequence.tbl}}{ a data.frame with the OTU sequences (see \code{\link{readFasta}}) and taxonomy, if supplied.}
}
}
\description{
Creating a list containing data from microbial community profiling at MiDiv.
}
\details{
This function stores the data structure from the processing of microbial community
sequencing data in a \code{list} that we refer to as a *midiv-object*.

The first four arguments are either data structure already read into R, or names of the files to read.

The \code{metadata.tbl} is a data.frame with one row for each sample, containing metadata
for each sample in the columns. It must have one column that uniquely identifies each sample.
This is specified in \code{sample_id_column} and is by default \code{"SampleID"}.
If a file name is specified, it is assumed to be a tab-delimited text file.

The \code{readcount.mat} is a *matrix* with readcounts, one row for each OTU and
one column for each sample. If a file name is specified, it is assumed to be
a tab-delimited text file, with OTUs in the rows and the samples in the columns,
but where the first column contains the OTU identifying texts (typically OTU1, OTU2,...).

The \code{sequence.tbl} is a Fasta-table with centroid sequences, see \code{\link{microseq::readFasta}},
or the name of a FASTA-file. These sequences are the centroid sequences for
each OTU, and the texts in the Header column must match the texts identifying the
OTUs in the \code{readcount.mat} above (typically OTU1, OTU2,...).

The \code{taxonomy.tbl} may be supplied, and must then be a table where the first
column is named \code{OTU} and contains the texts identifying the OTUs (typically OTU1, OTU2,...).
The remaining columns should list the taxonomy at various ranks,
and columns with (SINTAX) confidence scores are also allowed. If \code{taxonomy.tbl} is a file name
the file must be a tab-delimited text file with the columns as described above.
NB! In the created \code{midiv} object this is merged with the \code{sequence.tbl}.

The argument \code{filter_samples} is only used if the samples in the \code{metadata.tbl}
and \code{readcount.mat} are not the same. If \code{filter_sample = "metadata"} the samples
in this table are kept, and the \code{readcount.mat} is trimmed accordingly.

The argument \code{filter_OTUs} is only used if the OTUs in the \code{readcount.mat} and
\code{sequence.tbl} are not the same. If \code{filter_OTUs = "sequence"} the samples
in this table are kept, and the \code{readcount.mat} is trimmed/extended accordingly.
}
\author{
Lars Snipen.
}
