% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPS.R
\name{openPS}
\alias{openPS}
\title{Creating a phyloseq-like object}
\usage{
openPS(
  metadata.file,
  readcounts.file,
  centroids.file,
  taxonomy.file = NULL,
  sintax.threshold = 0
)
}
\arguments{
\item{metadata.file}{Full name of text-file with sample metadata.}

\item{readcounts.file}{Full name of text-file with readcounts data, assume samples in the columns.}

\item{centroids.file}{Full name of fasta-file with OTU sequences.}

\item{taxonomy.file}{Full name of text-file with the OTU taxonomy table results (optional).}

\item{sintax.threshold}{A value between 0.0 and 1.0 indicating confidence threshold for the taxonomy (optional).}
}
\value{
A list with the elements:

\itemize{
  \item{\code{sample_data.tbl}}{ a data.frame with one row for each sample. Its rownames are the SampleID's.}
  \item{\code{otu_table.mat}}{ a matrix with the readcounts. These may later be transformed. The samples are in the columns, the OTUs in the rows.}
  \item{\code{sequence.tbl}}{ a data.frame with the OTU sequences (see \code{\link{readFasta}}).}
  \item{\code{tax_table.mat}}{ a matrix with the taxonomy, one row for each OTU.}
}
}
\description{
Creating an object similar to a phyloseq-object, but as an open list.
}
\details{
This function reads the input files, and store their content in a
\code{list} that we refer to as an *openPS object*.

If a \code{taxonomy.file} is specified, the \code{sintax.threshold} will be used to assign
all classification with a score below this threshold to \code{"unclassified"} in the \code{tax_table.mat}.
}
\author{
Lars Snipen.
}
