% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefying.R
\name{rarefaction_curves}
\alias{rarefaction_curves}
\title{Computes rarefaction curves}
\usage{
rarefaction_curves(phyloseq.object, step = 100, plot = TRUE)
}
\arguments{
\item{phyloseq.object}{A phyloseq object.}

\item{step}{The step-size used to choose different down-sampling levels (readcounts).}

\item{plot}{Logical, plotting the curves or return the table for later plotting.}
}
\value{
If \code{plot = TRUE} a \code{ggplot} object, if \code{plot = FALSE}
the table with the data for plotting the curves. The latter is useful if you want
a different display of the curves
}
\description{
Computes rarefaction curves for a phyloseq object.
}
\details{
This function takes as input a phyloseq object, and uses the function
\code{\link{rarecurve}} in the \code{vegan} package for the computations.
}
\author{
Lars Snipen.
}
